Add multiplayer component for displaying stats
