+++ client/src/components/Multiplayer/api.js
@@ -1,0 +1,5 @@
+const sendMessage = (socket, message) => {
+  socket.emit('newClientMessage', message);
+};
+
+module.exports = sendMessage;
